refactor(v010): carve voicegw dashboard out of _legacy
Tenth per-command carve-out in section 2.

Adds voicegateway/cli/dashboard.py owning the 'dashboard' command.
Body unchanged from v0.0.5; lazy imports of uvicorn and
dashboard.api.main stay inside the function so the cli does not
eagerly load the dashboard extras for unrelated subcommands.

13 commands still register on the shared app; pytest tests/test_cli.py
46/46 pass; ruff and mypy clean across the cli package (11 source
files now).
