+++ voicegateway/cli/__init__.py
@@ -26,6 +26,7 @@ from __future__ import annotations
 # shared ``app`` via ``@app.command(...)`` decorators that run at
 # import time. Order does not matter; commands carry their own names.
 from voicegateway.cli import costs as _costs  # noqa: F401, E402
+from voicegateway.cli import dashboard as _dashboard  # noqa: F401, E402
 from voicegateway.cli import init as _init  # noqa: F401, E402
 from voicegateway.cli import logs as _logs  # noqa: F401, E402
 from voicegateway.cli import projects as _projects  # noqa: F401, E402
