+++ voicegateway/cli/dashboard.py
@@ -1,0 +1,44 @@
+"""``voicegw dashboard`` command.
+
+Carved out of voicegateway/cli/_legacy.py during the v0.1.0 section-2
+refactor. Boots the FastAPI dashboard backend (``dashboard.api.main``)
+under uvicorn against the loaded gateway. Defaults to 0.0.0.0:9090
+which is the v0.0.5 contract.
+"""
+
+from __future__ import annotations
+
+import typer
+
+# See voicegateway/cli/init.py for the rationale on importing
+# ``app`` and ``console`` from ``_legacy`` rather than from the
+# package during the v0.1.0 migration period.
+from voicegateway.cli._legacy import _load_gateway, app, console
+
+
+@app.command(name="dashboard")
+def dashboard_cmd(
+    config: str = typer.Option(None, "--config", "-c", help="Path to voicegw.yaml"),
+    host: str = typer.Option("0.0.0.0", "--host", help="Dashboard host"),
+    port: int = typer.Option(9090, "--port", help="Dashboard port"),
+) -> None:
+    """Start the web dashboard."""
+    try:
+        import uvicorn
+    except ImportError as e:
+        console.print(
+            "[red]Dashboard dependencies not installed. "
+            "Run: pip install 'voicegateway[dashboard]'[/red]"
+        )
+        raise typer.Exit(1) from e
+
+    gw = _load_gateway(config)
+    from voicegateway.core.auth import describe_auth, load_api_keys
+
+    console.print(f"[green]VoiceGateway dashboard at http://{host}:{port}[/green]")
+    console.print(f"[cyan]{describe_auth(load_api_keys(gw.config.auth))}[/cyan]")
+
+    import dashboard.api.main as dashboard_app
+
+    dashboard_app.configure(gw)
+    uvicorn.run(dashboard_app.app, host=host, port=port)
